test(functional): attempt to write single function for endpoints tests
