--- tests/functional/endpoints/dualstack/index.spec.ts
@@ -1,22 +1,0 @@
-import { join } from "path";
-
-import testCases from "./test_cases_supported.json";
-
-const getClientPackageName = (sdkId: string) =>
-  `client-${sdkId
-    .split(" ")
-    .map((word) => word.toLowerCase())
-    .join("-")}`;
-
-describe("endpoints.dualstack", () => {
-  for (const { sdkId, region, hostname } of testCases) {
-    const clientPackageName = getClientPackageName(sdkId);
-    it(`testing "${clientPackageName}" with region: ${region}`, async () => {
-      const { defaultRegionInfoProvider } = await import(
-        join("..", "..", "..", "..", "clients", clientPackageName, "src", "endpoints")
-      );
-      const regionInfo = await defaultRegionInfoProvider(region, { isDualstackEndpoint: true });
-      expect(regionInfo.hostname).toEqual(hostname);
-    });
-  }
-});
