--- tests/functional/endpoints/fips/test_cases_unsupported.json
@@ -1,26 +1,0 @@
-[
-  {
-    "endpointPrefix": "iotsecuredtunneling",
-    "sdkId": "IoTSecureTunneling",
-    "region": "ap-east-1",
-    "hostname": "api.tunneling.iot-fips.ap-east-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "iotsecuredtunneling",
-    "sdkId": "IoTSecureTunneling",
-    "region": "ca-central-1",
-    "hostname": "api.tunneling.iot-fips.ca-central-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "iotsecuredtunneling",
-    "sdkId": "IoTSecureTunneling",
-    "region": "cn-north-1",
-    "hostname": "iotsecuredtunneling-fips.cn-north-1.amazonaws.com.cn"
-  },
-  {
-    "endpointPrefix": "iotsecuredtunneling",
-    "sdkId": "IoTSecureTunneling",
-    "region": "us-gov-east-1",
-    "hostname": "api.tunneling.iot-fips.us-gov-east-1.amazonaws.com"
-  }
-]
